Application push to AWS. Edited config files for AWS use
