+++ .ebextensions/01_python.config
@@ -1,0 +1,5 @@
+option_settings:
+    aws:elasticbeanstalk:application:environment:
+        PYTHONPATH: "/var/app/current:$PYTHONPATH"
+    aws:elasticbeanstalk:container:python:
+        WSGIPath: "application:application"
