+++ .ebextensions/02_packages.config
@@ -1,0 +1,5 @@
+packages:
+  yum:
+    python3-devel: []
+    mariadb-devel: []
+    gcc: []
