+++ .ebextensions/03_timeout.config
@@ -1,0 +1,4 @@
+ option_settings:
+   - namespace: aws:elb:policies
+     option_name: ConnectionSettingIdleTimeout
+     value: 300
