--- .ebextensions/python.config
@@ -1,3 +1,0 @@
-option_settings:
-aws:elasticbeanstalk:container:python:
-WSGIPath: application:application
