+++ app/templates/calendar.html
@@ -39,6 +39,7 @@
             </div>
     </div>
     <div class="col mt-5 ms-5 me-5">
+        <div id="chart-wrapper">
                 <canvas id="chart" class="card border-success"></canvas>
                     <script>
