+++ app/templates/search.html
@@ -28,7 +28,7 @@
                                     <input type="radio" name="typeValue" id="Psuedocereal" value="Psuedocereal">Psuedocereal</input><br>
                                     <input type="radio" name="typeValue" id="Herb" value="Herb">Herb</input><br>
                                     <input type="radio" name="typeValue" id="Vegetable" value="Vegetable">Vegetable</input><br>
-                                    <input type="radio" name="typeValue" id="Bean" value="Bean">Bean</input><br>
+                                    <input type="radio" name="typeValue" id="Bean" value="Legume">Bean</input><br>
                                     <input type="radio" name="typeValue" id="Root" value="Root">Root</input><br>
                                     <input type="radio" name="typeValue" id="Fruit" value="Fruit">Fruit</input>
                                 </ul>
@@ -78,6 +78,7 @@
                                     <b> Family Name: </b>{{plantnames.family_name}}<br>
                                      <b>Plant Type: </b>{{plantnames.plant_type }}<br>
                                     <b>Life Cycle: </b>{{plantnames.life_cycle }}<br>
+                                    <b>Preferred Season: </b>{{plantnames.pref_season}}
                                 </div>
                             </div>
                                 {% else %}
