--- blog.py
@@ -1,3 +1,0 @@
-from app import app
-
-app.run(debug=True)
