+++ dbconfig.py
@@ -1,7 +1,4 @@
-from sqlalchemy import create_engine, text
-from sqlalchemy.orm import sessionmaker
-from flask import Flask
-from flaskext.mysql import MySQL
+from sqlalchemy import create_engine
 
 # instance identifier - plant-index-database
 # username - admin
